--- trade/src/main/java/org/coin/trade/pipeline/loop/AsyncLoop.java
@@ -1,6 +1,0 @@
-package org.coin.trade.pipeline.loop;
-
-public interface AsyncLoop {
-    void runAsyncLoop(int count);
-    void stopAsyncLoop();
-}
