--- trade/src/main/java/org/coin/trade/pipeline/reader/ItemReader.java
@@ -1,8 +1,0 @@
-package org.coin.trade.pipeline.reader;
-
-/**
- * @param <I> type to be read
- */
-public interface ItemReader<I> {
-    I read();
-}
