--- trade/src/main/java/org/coin/trade/pipeline/writer/ItemWriter.java
@@ -1,5 +1,0 @@
-package org.coin.trade.pipeline.writer;
-
-public interface ItemWriter <O>{
-    void write(O o);
-}
