+++ trade/src/main/java/org/coin/trade/queue/PipelineReaderBlockingQueue.java
@@ -1,7 +1,7 @@
 package org.coin.trade.queue;
 
 import org.coin.price.queue.MessageQueue;
-import org.coin.trade.dto.pipeline.reader.ReadOrderDto;
+import org.coin.trade.dto.pipeline.async.reader.ReadOrderDto;
 import org.springframework.stereotype.Component;
 
 import java.util.concurrent.BlockingQueue;
