+++ trade/src/main/java/org/coin/trade/queue/ProcessedOrderMessageBlockingQueue.java
@@ -1,7 +1,7 @@
 package org.coin.trade.queue;
 
 import org.coin.price.queue.MessageQueue;
-import org.coin.trade.dto.pipeline.writer.WriteOrderDto;
+import org.coin.trade.dto.pipeline.async.writer.WriteOrderDto;
 import org.springframework.stereotype.Component;
 
 import java.util.concurrent.LinkedBlockingQueue;
