+++ user/src/main/java/org/coin/wallet/repository/WalletBulkUpdateRepository.java
@@ -2,7 +2,7 @@ package org.coin.wallet.repository;
 
 import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
-import org.coin.trade.dto.pipeline.writer.ProcessedOrderDto;
+import org.coin.trade.dto.pipeline.async.writer.ProcessedOrderDto;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
